requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0
httpx>=0.25.0
httpx-sse>=0.4.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
aiofiles>=23.2.0